refactor: use custom hook to log in
refactor: use custom hook to log in
